graduation/毕设4-cursor/requirements.txt
# 基础科学计算
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 计算机视觉
opencv-python>=4.5.0
open3d>=0.13.0
pyrealsense2>=2.50.0
pillow>=8.0.0

# 机器学习
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0

# 数据处理
pyyaml>=5.4.0
h5py>=3.6.0

# 工具包
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyserial>=3.5

# 性能优化
numba>=0.54.0
psutil>=5.8.0

# 日志和监控
tensorboard>=2.7.0

# 测试
pytest>=6.2.0
pytest-cov>=2.12.0

# 开发工具
black>=21.7b0
flake8>=3.9.0
mypy>=0.910 